refactor: restructure app
