+++ src/components/ErrorBoundary/ErrorBoundary.module.css
@@ -1,0 +1,8 @@
+.errorpage {
+  min-height: 100vh;
+  display: grid;
+  place-content: center;
+
+  font-size: 3rem;
+  color: red;
+}
