+++ src/components/ErrorButton/ErrorButton.module.css
@@ -8,12 +8,3 @@
   font-weight: 700;
   color: red;
 }
-
-.errorpage {
-  min-height: 100vh;
-  display: grid;
-  place-content: center;
-
-  font-size: 3rem;
-  color: red;
-}
