+++ src/components/search/Search.tsx
@@ -1,4 +1,4 @@
-import styles from './search.module.css';
+import styles from './Search.module.css';
 
 import { Component, createRef } from 'react';
