+++ src/components/spinner/Spinner.tsx
@@ -1,4 +1,4 @@
-import styles from './spinner.module.css';
+import styles from './Spinner.module.css';
 
 function Spinner() {
   return (
