+++ src/main.tsx
@@ -1,8 +1,9 @@
+import './index.css';
+
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
-import './index.css';
-import { ErrorBoundary } from './components/error/ErrorBoundary.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
